CLN: Split _find_gridprop_in_eclrun
Parts of _find_gridprop_in_eclrun split moved into _gridprop_import_eclrun
